JP - Huge linter update. I spent a ton of time getting both `tslint` and `prettier` working. `prettier` does some pretty formatting, and `tslint` takes care of code related linting issues (they sort of overlap). I also found a program that checks to make sure they're running together
On-top of setting up the linter I also spent time getting rid of all the linter errors. The linter(s) can fix some errors and stuff automatically, but not others. All the other linting issue I had to manually correct.

Since I did spend time correcting all the linting issues, this is a huge update. It touches almost every piece of code. The nice thing is that the code is now much cleaner and easier to work with.