Simplified logo loading and added some TODOs
I didn't like the logo loading hack that what specifically check for a file in some beta branch. I thought it woulb be easier to just simply upload the png file to the master assets folder. It won't hurt anything if the png file is there and nothing uses it. Then I referenced the assets folder via its root (/assets/4DVDLogoV1.png) and that took care of it loading fine in any branch we work on in the future. I also added some TODOs around some hacks that are in place to rename the folders because we should simply be getting those names from the DB instead.